+++ spec/integration/config.spec.js
@@ -9,6 +9,7 @@ var config = {
 	defaultUser: true,
 };
 var get;
+var harnessFn = require( '../../src/harness' );
 
 describe( 'Configuration', function() {
 	var harness, cwd;
@@ -35,7 +36,7 @@ describe( 'Configuration', function() {
 				var testConfig = _.extend( {}, config, {
 					static: false
 				} );
-				harness = require( './harness.js' )( testConfig );
+				harness = harnessFn( testConfig );
 				harness.start();
 			} );
 
@@ -65,7 +66,7 @@ describe( 'Configuration', function() {
 
 			before( function() {
 				cwd = process.cwd();
-				harness = require( './harness.js' )( _.extend( {}, config ) );
+				harness = harnessFn( _.extend( {}, config ) );
 				harness.start();
 			} );
 
@@ -95,7 +96,7 @@ describe( 'Configuration', function() {
 				var testConfig = _.extend( {}, config, {
 					static: './public/txt'
 				} );
-				harness = require( './harness.js' )( testConfig );
+				harness = harnessFn( testConfig );
 				harness.start();
 			} );
 
@@ -131,7 +132,7 @@ describe( 'Configuration', function() {
 						}
 					}
 				} );
-				harness = require( './harness.js' )( testConfig );
+				harness = harnessFn( testConfig );
 				harness.start();
 			} );